Refactor expense and spreadsheet services to enhance functionality and improve type definitions. Added support for multiple file uploads in the expense service, updated related types, and introduced new utility functions for spreadsheet processing. Enhanced test coverage for new functionalities.
