+++ packages/business/src/finance/bill/business/spreadsheet/index.ts
@@ -1,0 +1,2 @@
+export { default as BillSpreadsheetBusiness } from './spreadsheet';
+export * from './types';
