+++ packages/business/src/finance/expense/index.ts
@@ -1,5 +1,5 @@
 export { default as Expense } from './expense';
 export { ExpenseService } from './service';
-export { ExpenseBusiness, type AllExpensesCalculated } from './business';
+export { ExpenseBusiness, type PrepareForCreationParams, type PrepareForCreationResult } from './business';
 export * from './types';
 export * from './enum';
