+++ packages/business/src/finance/month/business/business.ts
@@ -16,10 +16,8 @@ type GeneratePersistListMonthParams = GeneratePersistMonthParams & {
     months?: Array<PersistMonthParams>;
 };
 
-type GenerateMonthListCreationParams = GeneratePersistListMonthParams;
-
 export default class MonthBusiness {
-    public generateMonthListUpdateParameters(months?: Array<Month>, monthsToPersist?: Array<PersistMonthParams>): Array<PersistMonthParams> | undefined {
+    public generateMonthListUpdateParameters(months?: Array<Month>, monthsToPersist?: Array<PersistMonthParams>, paid?: boolean): Array<PersistMonthParams> | undefined {
         if ((months && months?.length > 0) && (monthsToPersist && monthsToPersist?.length > 0)) {
             const monthList = monthsToPersist?.filter((m) => {
                 if (!m.code && !m.month) {
@@ -35,7 +33,10 @@ export default class MonthBusiness {
             if (monthList.length <= 0) {
                 return;
             }
-            return monthList;
+            return monthList.map((month) => ({
+                ...month,
+                paid: paid !== undefined ? paid : month.paid
+            }));
         }
         return;
 
@@ -71,7 +72,7 @@ export default class MonthBusiness {
         return [this.generatePersistMonthParams(params)];
     }
 
-    public generateMonthListCreationParameters(params: GenerateMonthListCreationParams): Array<PersistMonthParams> {
+    public generateMonthListCreationParameters(params: GeneratePersistListMonthParams): Array<PersistMonthParams> {
         const months = this.generatePersistListMonthParams(params);
         return MONTHS.map((month) => {
             const code = getCurrentMonthNumber(month);
